requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
pymysql==1.1.1
requests==2.32.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
Pillow==10.4.0
